Committing my patch from #IO-101, fixing an <int> overrun in readSwappedLong. Many thanks to Jose Pinto for finding this
git-svn-id: https://svn.apache.org/repos/asf/jakarta/commons/proper/io/trunk@482841 13f79535-47bb-0310-9956-ffa450edef68

+++ src/java/org/apache/commons/io/EndianUtils.java
@@ -218,12 +218,12 @@ public class EndianUtils {
             ( ( data[ offset + 0 ] & 0xff ) << 0 ) +
             ( ( data[ offset + 1 ] & 0xff ) << 8 ) +
             ( ( data[ offset + 2 ] & 0xff ) << 16 ) +
-            ( ( data[ offset + 3 ] & 0xff ) << 24 );
+            ( ( (long) ( data[ offset + 3 ] & 0xff ) ) << 24 );
         long high = 
             ( ( data[ offset + 4 ] & 0xff ) << 0 ) +
             ( ( data[ offset + 5 ] & 0xff ) << 8 ) +
             ( ( data[ offset + 6 ] & 0xff ) << 16 ) +
-            ( ( data[ offset + 7 ] & 0xff ) << 24 );
+            ( ( (long) ( data[ offset + 7 ] & 0xff ) ) << 24 );
         return low + (high << 32);
     }
